+++ docs/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- catch TypeError exceptions from BeautifulSoup in HtmlTagFilter
+
 ## [2.4.0] - 2022-04-05
 
 ### Added